Move NamedTuples to types.py and rename SearchParams to ExtractedSearchParams
