+++ src/crewai/rag/chromadb/client.py
@@ -2,7 +2,7 @@
 
 import hashlib
 from collections.abc import Mapping
-from typing import Any, NamedTuple, TypeGuard
+from typing import Any, TypeGuard
 
 from chromadb.api import AsyncClientAPI, ClientAPI
 from chromadb.api.configuration import CollectionConfigurationInterface
@@ -20,7 +20,11 @@ from chromadb.api.types import (
 )
 from typing_extensions import Unpack
 
-from crewai.rag.chromadb.types import ChromaDBClientType
+from crewai.rag.chromadb.types import (
+    ChromaDBClientType,
+    PreparedDocuments,
+    ExtractedSearchParams,
+)
 from crewai.rag.core.base_client import (
     BaseClient,
     BaseCollectionParams,
@@ -55,44 +59,6 @@ def _is_async_client(client: ChromaDBClientType) -> TypeGuard[AsyncClientAPI]:
         True if the client is an AsyncClientAPI, False otherwise.
     """
     return isinstance(client, AsyncClientAPI)
-
-
-class PreparedDocuments(NamedTuple):
-    """Prepared documents ready for ChromaDB insertion.
-
-    Attributes:
-        ids: List of document IDs
-        texts: List of document texts
-        metadatas: List of document metadata mappings
-    """
-
-    ids: list[str]
-    texts: list[str]
-    metadatas: list[Mapping[str, str | int | float | bool]]
-
-
-class SearchParams(NamedTuple):
-    """Extracted search parameters for ChromaDB queries.
-
-    Attributes:
-        collection_name: Name of the collection to search
-        query: Search query text
-        limit: Maximum number of results
-        metadata_filter: Optional metadata filter
-        score_threshold: Optional minimum similarity score
-        where: Optional ChromaDB where clause
-        where_document: Optional ChromaDB document filter
-        include: Fields to include in results
-    """
-
-    collection_name: str
-    query: str
-    limit: int
-    metadata_filter: dict[str, Any] | None
-    score_threshold: float | None
-    where: Where | None
-    where_document: WhereDocument | None
-    include: Include
 
 
 class ChromaDBCollectionCreateParams(BaseCollectionParams, total=False):
@@ -158,16 +124,16 @@ def _prepare_documents_for_chromadb(
 
 def _extract_search_params(
     kwargs: ChromaDBCollectionSearchParams,
-) -> SearchParams:
+) -> ExtractedSearchParams:
     """Extract search parameters from kwargs.
 
     Args:
         kwargs: Keyword arguments containing search parameters.
 
     Returns:
-        SearchParams with all extracted parameters.
+        ExtractedSearchParams with all extracted parameters.
     """
-    return SearchParams(
+    return ExtractedSearchParams(
         collection_name=kwargs["collection_name"],
         query=kwargs["query"],
         limit=kwargs.get("limit", 10),
